rip out PromiseWatcher logic
There was a race condition. The events were as follows:

- weakPromise is completed. This schedules, in a batched fashion, a function which completes of the upstream promise.
- the weakPromise is GC'd
- The weakPromise watcher runs, and detects that the weakPromise is GC'd, but the upstream promise is not yet completed (because there is a batch-scheduled function pending)
- The batch-scheduled function is scheduled

I considered that it was possible that the
strongPromise.completeWith(weakPromise.Future) was not causing the
weakPromise to be strongly held, but this seems very unlikely upon
reading the code; the weakPromise has some state to be modified once it
gets completed, so I assume the garbage collection event is happening
AFTER the weakPromise is being completed, and the strongPromise
completing Function0 is scheduled in the batched execution context.